requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
SQLAlchemy==1.4.25
psycopg2==2.9.1
python-multipart==0.0.5
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.3.0